Added BOOKMARK Comments                                     [2013-02-27]
I recently added the 'addons' plugin to Geany, the editor I use on
Linux, and it contains a very handy feature to scan sources for tags
like FIXME:, BOOKMARK: and the like. Custom tags can also be defined.

So I added some major bookmark locations to easily navigate around
multiple sources. What's also nice is that of course these tags can be
listed with grep. For instance, grep -r -n 'BOOKMARK:' *.ASM will
display a list of all bookmark comments with line-numbers in the
assembler sources in the current and underlying directories.

Changes:
  o Started fixing the CRC issue
    Since AiR-BOOT now updates the PBR on HPFS and JFS, the CRC needs to
    be recalculated and stored in the IPT. Otherwise AB will think the
    PBR is infected and a virus message will appear.
  o Changed PART_UpdateBootRecordCRC
    Only PartitionSector could be CRC'd because it was hard coded.
    Also, SI got immediately overwritten and that might have been an
    unsurfaced bug, even in v.106. Have to investigate this further.
    Changed to use BX as a pointer to the sector to be CRC'd.

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@52 3beef77e-8972-4c43-b739-0c5365d60c6d