Preparing for v1.1.0                                        [2013-04-05]
During v1.0.8 it was decided that odd minor numbers will be
test-versions. Since v1.0.8 was actually a test-version with several
test-releases but with an even minor number, it was decided to bump the
final release to v1.1.0.

Thus, v1.1.1 will be the next test-version after the upcoming official
v1.1.0 release. When such test-versions are published, for instance to
get feedback, or solve a particular issue, the build-date and BLDLEVEL
information will identify such a test-release. This information can be
viewed by pressing the TAB-key while the Main Menu is displayed. Such
test-releases will also have a slightly different visual appearance to
distinguish them from official releases.

Because all test-versions between official releases identify themselves
with the same version number, upgrading from one test-release to another
requires the /forcecode option on the Installer.

Following this scheme, v1.1.2 will be the next official release after
the upcoming v1.1.0 release.

It is possible however, that major changes break compatibility with
previous releases, in which case the middle number will be incremented
to reflect this.

Note that this commit still identifies as v1.0.8.

Changes
  o Bumped version information to v1.1.0
  o Updated version dependent code in the Installers and SETABOOT.EXE

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@54 3beef77e-8972-4c43-b739-0c5365d60c6d