AiR-BOOT v1.1.0-RELEASE                                     [2013-04-05]
This is the version as included on the eCS v2.2-beta CD's.
It was technically a pre-v1.1.0 release, with the final v1.1.0,
including some minor patches, to be included with eCS v2.2-GA.
Because this version is now known to be v1.1.0, we'll keep it that way.

The minor patches and meanwhile done fixes will soon appear as v1.1.1
test-version commits to be consolidated to a v1.1.2 release.
Issues regarding removable media are planned for the v1.1.4 release as
this requires removal of obsolete code and optimizing existing code to
create space in the ~32K block (track0) the loader resides in.

With this version of AiR-BOOT, the display of the version number has
changed to be compatible with the WarpIN versioning system.
So all digits are now separated by a dot. This change is only visual and
tools relying on the way AiR-BOOT stores it's version numer internally
are not affected.

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@55 3beef77e-8972-4c43-b739-0c5365d60c6d