AiR-BOOT v1.1.1-TESTING   !! MASSIVE LOWERCASING UPCOMING !!   [opening]
  !! Be sure to have a clean workdir before running 'svn update' !!

The DOS era has well passed and the UPPERCASE source-file names are
really starting hurt my eyes. Besides that, AiR-BOOT can also be built
under Linux, where the file-systems are mostly case-sensitive with
lowercase being the standard. So it's time to grab these sources and
hammer their case down, which is what will happen in the next commit.

Besides this lowercasing, some obsolete stuff will also be removed.

This v1.1.1-testing cycle will be relatively short with the main purpose
of integrating some patches and fixes from the local dev-queue. After
things look good enough a v1.1.2-RELEASE will be done.

Once on v1.1.3-testing, the plan is to also integrate the building of
the INF and PDF documentation into this project. That part will require
LFNs, but the building of core AiR-BOOT files, like the installers and
of course the bootloader itself will remain doable on 8.3 FAT systems.


git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@56 3beef77e-8972-4c43-b739-0c5365d60c6d