Get rid of the non-ASCII characters                     [v1.1.1-testing]
These characters use CP437 encoding, which causes problems when the
source is loaded into editors not supporting this encoding. This goes
especially for the '0adh' character which is part of the AiR-BOOT
configuration signature.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@69 3beef77e-8972-4c43-b739-0c5365d60c6d