Switching on serial port debugging                      [v1.1.1-testing]
Information is logged to the serial port and pressing some keys will
log some specific information. Press 'h' to see which keys are active.
Because of the initial info dump, it will take some time before the
menu appears.

If you build from this commit and do not have a serial port or have no
terminal attached, AiR-BOOT will most probably hang or take a long timr
before the menu is displayed.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@71 3beef77e-8972-4c43-b739-0c5365d60c6d