Temporarily disable checksum verification for code      [v1.1.1-testing]
In the development enviroment, we use VirtualBox with VMDK images.
To quickly test modifications, the AirBoot loader image is merged into
the VMDK image ouside of the VM. Because the 'airboot2.exe' installer
is not used, no checksum is embedded, so we temporarily disable
checksumming for the code section here.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@72 3beef77e-8972-4c43-b739-0c5365d60c6d