Allow CAD when halting system                           [v1.1.1-testing]
Of course halting with interrupts disabled is a safer halt than keeping
them enabled. This was probably done to minimize effects of (stealth)
boot-sector virusses of the time. Nowadays such beasts are rare and
being able to do a CAD after some error is definitely more convenient
than pressing the not-there-anymore hard reset switch or holding the
power button for several seconds.

Maybe we'll tie 'allow CAD when halted' to one of the anti-virus
setup bits. For now, system halts will allow CAD to do a soft reset.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@81 3beef77e-8972-4c43-b739-0c5365d60c6d