Added function to convert a byte to packed BCD          [v1.1.1-testing]
This enabled displaying small numbers (0>=n<=255) as decimal values.
The function returns a 16-bit value with the number of decimal digits
in the high nibble and the decimal digits in the remaining nibbles.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@87 3beef77e-8972-4c43-b739-0c5365d60c6d