LVM CRC-check was always forced to be successful        [v1.1.1-testing]
Wow, this is a naughty one !
Function 'DriveIO_LVMSectorValid' was introduced in 'DRIVEIO.ASM' in
r30: http://trac.netlabs.org/air-boot/changeset/30/#file19

The CRC-check was always forced to OK with 'stc' because it always
failed. This failing was the result of checking the CRC *before* the
'LVM_InitCRCTable' was called. So AirBoot never catched bad LVM CRCs!

The LVM stuff is now being revisited, but we'll correct it here for the
record.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@99 3beef77e-8972-4c43-b739-0c5365d60c6d