Prevent LVM CRC-check indicating OK on zero sectors     [v1.1.1-testing]
When 'LVM_InitCRCTable' is not called yet, an LVM CRC-check on a sector
with all zeros would indicate a valid LVM CRC. Eventhough this is not
the case _after_ 'LVM_InitCRCTable' has been called, we rule out this
situation to guard agains coding bugs.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@101 3beef77e-8972-4c43-b739-0c5365d60c6d