Changed the layout of the BSS to make space             [v1.1.1-testing]
We are going to store the complete disk information data from the
INT13 and INT13X BIOS calls (08h and 48h). INT13X v3+ can return 42h
of disk information. A total of 64 disks are supported, so some more
space in the BSS was needed. This was done by moving the BSS downwards
by 2000h (8K) bytes. While function 48h only returns 42h bytes, we
allocate 80h bytes per disk, so the size of the block is a multiple
of 2, so indexing can be done using bit shifts.

Data blocks containing related information are clustered and aligned
on 16 byte boundaries. That gives us neat offsets with the lower 4 bits
zero.

Supporting 64 disks might seem a bit overblown, especially since
AirBoot only supports a maximum of 45 partitions across all disks.
It is mainly done with regard to UEFI and future enhancements in mind.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@112 3beef77e-8972-4c43-b739-0c5365d60c6d