Force 'write-verify' off on sector writing              [v1.1.1-testing]
In INT13X v1.0 and v2.0, AL.0 was used to indicate 'verify write'.
In v2.1+ this seems to have moved to AL.1 -- wonder why that change
was needed.

Anyway, to ensure compatibility and prevent coding mistakes this
feature is turned off when writing sectors.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@113 3beef77e-8972-4c43-b739-0c5365d60c6d