Improved method of locating the Master LVM-sector       [v1.1.1-testing]
This potentially fixes 'LOAD ERROR' issues with USB sticks present.

The old method assumed the fixed locations 255,127 and 63, expressed
as non-LBA sectors but converted to LBA when doing the sector loads.
The above locations are *OS/2 Geometry* and have actually no relation
to how the BIOS views the geometery of the disk.

Then there was also a bug, where the 63 value was overwritten with the
BIOS SPT queried using INT13X, to try that value in case none of the
above had a valid LVM-sector. To make a long story short: this could
go very bad when USB sticks were present at boot time or with disks
present that use OS/2 extended geometry and normal geometry in some
particular order.

The new method uses LBA addressing to scan from LBA sector 254 downwards
and, when a valid LVM-sector is found, also veryfies it is found at the
LBA location it should be, which is derived from the OS/2 geometry of
the LVM-sector itself.

Note that this new method is not used in this commit yet, but will be
integrated in upcoming commits that improve more disk handling aspects.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@115 3beef77e-8972-4c43-b739-0c5365d60c6d