Added a function that specifically loads the MBR        [v1.1.1-testing]
Besides loading the MBR, this function also checks if a signature is
present, if primary and/or extended partitions are present, and also
if AirBoot is installed on the disk.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@126 3beef77e-8972-4c43-b739-0c5365d60c6d