New storage method for disk information                 [v1.1.1-testing]
Having separate areas for BIOS, INT13(X) and LVM information makes
addressing them a bit tedious. Now all disk information is going to be
clustered in one structure with an instance of that structure for each
disk. The disk information will be gathered during the scanning phase
so it can be easily accessed afterwards.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@128 3beef77e-8972-4c43-b739-0c5365d60c6d