Added method to force build failure                     [v1.1.1-testing]
Usually, commits are buildable and some minor checking has been done
that whole disks do not get wiped. However, there are situations where
stuff is in such a flux that it would be better to do non-buildable
commits. That makes it possible to record the changes piecemeal without
the danger of accidently using the resulting loader. Of course everyone
is clever enough to _never_ build commits marked as 'testing' and
install the result in environments where wiping all attached disks and
sticks would be a problem...

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@132 3beef77e-8972-4c43-b739-0c5365d60c6d