Reimplemented loading of the Master LVM sector          [v1.1.1-testing]
The new method now uses the disk information gathered in an earlier
stage and just loads an already found Master LVM sector. This fix gets
rid of the ugly and faulty 'sec_per_track_table' way of doing it.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@136 3beef77e-8972-4c43-b739-0c5365d60c6d