Removed all the disk stuff from 'PRECRAP'               [v1.1.1-testing]
This stuff has always been in the wrong place because no scanning has
taken place yet. The stuff in 'PRECRAP' should be concerned with
initializing stuff only.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@139 3beef77e-8972-4c43-b739-0c5365d60c6d