Added function to scan all disks for base information   [v1.1.1-testing]
This function does the initial scanning of disks with the purpose of
populating the DISKINFO structures for each disk.

It precedes the second scanning stage, which involves scanning disks to
gather info about partitions etc.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@143 3beef77e-8972-4c43-b739-0c5365d60c6d