Fixed bug for new LVM sector load method                [v1.1.1-testing]
Getting the LVM_SPT was done before DL was loaded with the correct
BIOS disk number. Fixed.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@149 3beef77e-8972-4c43-b739-0c5365d60c6d