Prepared 'LVM_GetDriveLetter' for reimplementation      [v1.1.1-testing]
This is one ugly mthrfscker of which we won't even record the diff with
its reimplementation.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@153 3beef77e-8972-4c43-b739-0c5365d60c6d