Splitted-up the load/save LVM sector functions          [v1.1.1-testing]
The high-level variants use '[CurPartition_Location]' to get the LBA
address of the MBR/EBR. However, we also need access to LVM information
at other points and we don't want to mess with '[CurPartition_Location]'
there. So, the new low-level variants, *XBR, take the LBA address of
the MBR/EBR and BIOS disk number in registers. Since the IPT stores the
LBA address where the partition tables for a partition are located,
the *XBR funtions now provide a convenient way to load the corresponding
LVM sector.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@152 3beef77e-8972-4c43-b739-0c5365d60c6d