Preserve SI when using 'LVM_GetDriveLetter'             [v1.1.1-testing]
In several parts of the code, SI points to the IPT-entry of the
partition acted upon and is often assumed to be preserved across
helper functions.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@158 3beef77e-8972-4c43-b739-0c5365d60c6d