Set DEBUG_LEVEL=2 for building -- catch JWasm SIZE bug  [v1.1.1-testing]
With -W0, JWasm does not complain about stuff like 'mov [bx], 1'.
With -W4, JWasm does complain and states it assumes BYTE, and correct
code is generated. With 'add [bx], 1' however, assuming BYTE is stated
but WORD code is generated. This is a bug in JWasm v2.06d.
With DEBUG_LEVEL=2, such stuff is catched and flagged as an error.

Also removed the debug option from the linker flags, because they
produce an oversized image. They are not used anyway.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@160 3beef77e-8972-4c43-b739-0c5365d60c6d