Removed used assembler from build-info display          [v1.1.1-testing]
Building with the other assemblers is going to be removed to exploit
JWasm to the max. We've been using JWasm since revision #42, so nothing
new about that anymore. And we gain some precious code-space again.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@163 3beef77e-8972-4c43-b739-0c5365d60c6d