Fake write succes on removable disks                    [v1.1.1-testing]
Accessing, and thus writing, removable disks is highly dependent on
the BIOS USB mass-storage emulation. Also, some removable disks have
a small switch to set it to read-only. We don't want such stuff to halt
AiR-BOOT with a 'SAVE ERROR', so we fake success in such cases.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@168 3beef77e-8972-4c43-b739-0c5365d60c6d