Added a double check when writing LVM sectors           [v1.1.1-testing]
Writing an LVM sector is always the result of modifications to a
previously loaded one. So we first do a load and validate to make sure
the addressed sector contains LVM information.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@169 3beef77e-8972-4c43-b739-0c5365d60c6d